Move upsert-private-segment-list-modal to the new segments/modals directory
